chore: isolate streams types
