+++ packages/core/streams/index.ts
@@ -4,4 +4,6 @@ export * from './streaming-text-response'
 export * from './huggingface-stream'
 export * from './anthropic-stream'
 export * from './langchain-stream'
+
+export * from './types'
 export * from '../shared/types'
